Make details list work
after new folder/SB structure.
Next refactoring/organizing because of naming collisions.
