+++ src/stories/Blocks/material-manifestation-item/material-manifestation-item.scss
